+++ src/app/native/detail/planet-power/planet-power.component.ts
@@ -7,7 +7,7 @@ import {
 } from '@angular/core';
 import { NgStyle } from '@angular/common';
 import { IonicModule } from '@ionic/angular';
-import { Horoscope } from 'src/app/type/interface/response-data';
+import { Horoscope, ReturnHoroscope } from 'src/app/type/interface/response-data';
 import { PlanetName } from 'src/app/type/enum/planet';
 import { Horoconfig } from 'src/app/services/config/horo-config.service';
 import {
@@ -23,7 +23,7 @@ import {
   imports: [IonicModule, NgStyle],
 })
 export class PlanetPowerComponent implements OnInit, OnChanges {
-  @Input() horoscopeData: Horoscope | null = null;
+  @Input() horoscopeData: Horoscope | ReturnHoroscope | null = null;
 
   planetPowers: PlanetPower[] = [];