+++ src/app/native/detail/reception/reception.component.ts
@@ -6,7 +6,7 @@ import {
   SimpleChanges,
 } from '@angular/core';
 import { IonicModule } from '@ionic/angular';
-import { Horoscope } from 'src/app/type/interface/response-data';
+import { Horoscope, ReturnHoroscope } from 'src/app/type/interface/response-data';
 import { Horoconfig } from 'src/app/services/config/horo-config.service';
 import { degreeToDMS } from 'src/app/utils/horo-math/horo-math';
 import {
@@ -25,7 +25,7 @@ import {
   imports: [IonicModule],
 })
 export class ReceptionComponent implements OnInit, OnChanges {
-  @Input() horoscopeData: Horoscope | null = null;
+  @Input() horoscopeData: Horoscope | ReturnHoroscope | null = null;
   receptions: Reception[] = [];
   mutualReceptions: MutualReception[] = [];
   private initialized = false;